#937 Исправить все проигнорированные тесты
