+++ VkNet.Tests/Infrastructure/BaseTest.cs
@@ -93,16 +93,6 @@ namespace VkNet.Tests
 					State = "123456"
 				});
 
-			Mocker.Setup<ICaptchaHandler, string>(m => m.Perform(It.IsAny<Func<ulong?, string, string>>()))
-				.Callback(() => ReadJsonFile("Categories", "Messages", "GetById_NormalCase_Message"))
-				.Returns(() => Json);
-
-			Mocker.Setup<ICaptchaHandler, bool>(m => m.Perform(It.IsAny<Func<ulong?, string, bool>>()))
-				.Returns(true);
-
-			Mocker.Setup<ICaptchaHandler, int>(m => m.MaxCaptchaRecognitionCount)
-				.Returns(1);
-
 			Mocker.Setup<ICaptchaSolver, string>(m => m.Solve(It.IsAny<string>()))
 				.Returns("123456");
 
@@ -128,7 +118,7 @@ namespace VkNet.Tests
 			Api.RestClient = Mocker.Get<IRestClient>();
 			Api.NeedValidationHandler = Mocker.Get<INeedValidationHandler>();
 			Api.CaptchaSolver = Mocker.Get<ICaptchaSolver>();
-			Api.CaptchaHandler = Mocker.Get<ICaptchaHandler>();
+			SetupCaptchaHandler();
 
 			Api.Authorize(new ApiAuthParams
 			{
@@ -150,6 +140,10 @@ namespace VkNet.Tests
 		{
 			Json = null;
 			Url = null;
+		}
+
+		protected virtual void SetupCaptchaHandler()
+		{
 		}
 
 		protected VkResponse GetResponse()
